Add hugepage Downward API support.
In Kubernetes 1.20, an alpha feature was added to expose the requested
hugepages to the container via the Downward API. Add a new API to check
to see if the hugepage data is passed to the container via the Downward
API, and if so, return the data to the caller.

This is feature is disabled by default in Kubernetes. If enabled when
Kubernetes is deployed via FEATURE_GATES="DownwardAPIHugePages=true",
then Downward API values can be passed to the container.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>