Rename multusplugin as networkstatus
The multusplugin pkg was used to pull in the NetworkStatus map from the multus-cni
code base and parse the NetworkStatus annotation. Turns out that map was outdated
and the map that was being pushed in annotations was from the
network-attachment-definition-client library. So moved to pull from there and
remove references to Multus.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>