Add support for device-info-spec
Add support to app-netutil to implement the device-info-spec. Primarily, it consists
of pull additional data fron the DeviceInfo field in the NetworkStatus.

As part of supporting the device-info-spec, migrate app-netutil to use
network-attachment-definition-client data structs. app-netutil redefined some of
the data structs from Multus and Userspace CNI. Since
network-attachment-definition-client is now the location for these data structs,
migrate to app-netutil to use these instead of defining it's own.

Since app-netutil is not GA yet, the data structure returned by GetInterfaces()
changed. Once this package is moved to GA, the API will need to be backwards
compatible.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>