Move away from glog
glog requires `flag.parse()` to be called before logging. Because app-netutil
is a library to be consumed by other GO and C programs, and `flag.parse()`
consumes argv and argc, it is not desirable to call. So move away from glog.

First, move to a local log API so underlying logging can be swapped out.
Future change is to use another logging library.

Resolves: #36

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>