Move common C functions to utility folder
There were a few C functions for printing the retrieved data and
freeing the allocated memory that were duplicated between c_sample
and the dpdk-app-centos. Was originally trying to include the functions
in the cgo generated shared library. Gave up on that solution and
just moved them to a common directory and file so they could be
used in multiple places.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>